added protectin on update and add mutation, so they can be used only by admins with jwt token

+++ schema/schema.js
@@ -122,7 +122,9 @@ const Mutations = new GraphQLObjectType({
         projectUrl: { type: GraphQLString },
         images: { type: new GraphQLList(GraphQLString) },
       },
-      resolve(_parent, args) {
+      resolve(_parent, args, req) {
+        if (!req.user) throw new Error('Not an admin');
+
         const project = new Project(
           {
             title: args.title,
@@ -151,7 +153,9 @@ const Mutations = new GraphQLObjectType({
         projectUrl: { type: GraphQLString },
         images: { type: new GraphQLList(GraphQLString) },
       },
-      resolve(_parent, args) {
+      resolve(_parent, args, req) {
+        if (!req.user) throw new Error('Not an admin');
+
         return Project.findByIdAndUpdate(
           args.id,
           {
@@ -172,7 +176,9 @@ const Mutations = new GraphQLObjectType({
       args: {
         id: { type: new GraphQLNonNull(GraphQLID) },
       },
-      resolve(_parent, args) {
+      resolve(_parent, args, req) {
+        if (!req.user) throw new Error('Not an admin');
+
         return Project.findByIdAndDelete(args.id);
       },
     },
@@ -183,7 +189,9 @@ const Mutations = new GraphQLObjectType({
         type: { type: new GraphQLNonNull(GraphQLString) },
         image: { type: new GraphQLNonNull(GraphQLString) },
       },
-      resolve(_parent, args) {
+      resolve(_parent, args, req) {
+        if (!req.user) throw new Error('Not an admin');
+
         const tag = new Tag(
           {
             name: args.name,
@@ -204,7 +212,9 @@ const Mutations = new GraphQLObjectType({
         type: { type: GraphQLString },
         image: { type: GraphQLString },
       },
-      resolve(_parent, args) {
+      resolve(_parent, args, req) {
+        if (!req.user) throw new Error('Not an admin');
+
         return Tag.findByIdAndUpdate(args.id, {
           name: args.name,
           type: args.type,
@@ -217,7 +227,9 @@ const Mutations = new GraphQLObjectType({
       args: {
         id: { type: new GraphQLNonNull(GraphQLID) },
       },
-      resolve(_parent, args) {
+      resolve(_parent, args, req) {
+        if (!req.user) throw new Error('Not an admin');
+
         return Tag.findByIdAndDelete(args.id);
       },
     },
